Article and page backend

+++ backend/controllers/ArticleController.php
@@ -4,26 +4,44 @@ namespace backend\controllers;
 
 use Yii;
 use common\models\Article;
-use backend\models\search\ArticleSearch;
-use yii\web\Controller;
+use common\models\search\ArticleSearch;
+use common\components\web\Controller;
 use yii\web\NotFoundHttpException;
 use yii\filters\VerbFilter;
+use yii\helpers\Json;
+use yii\helpers\Url;
+use yii\helpers\ArrayHelper;
+use kartik\grid\EditableColumnAction;
 
 /**
  * ArticleController implements the CRUD actions for Article model.
  */
 class ArticleController extends Controller
 {
+    /**
+     * @inheritdoc
+     */
     public function behaviors()
     {
         return [
             'verbs' => [
                 'class' => VerbFilter::className(),
                 'actions' => [
-                    'delete' => ['post']
-                ]
-            ]
+                    'delete' => ['POST'],
+                ],
+            ],
         ];
+    }
+
+    public function actions()
+    {
+        return ArrayHelper::merge(parent::actions(), [
+            'edit' => [
+                'class' => EditableColumnAction::className(),
+                'modelClass' => Article::className(),
+                'showModelErrors' => true,
+            ]
+        ]);
     }
 
     /**
@@ -39,8 +57,28 @@ class ArticleController extends Controller
         ];
         return $this->render('index', [
             'searchModel' => $searchModel,
-            'dataProvider' => $dataProvider
+            'dataProvider' => $dataProvider,
         ]);
+    }
+
+    /**
+     * Displays a single Article model.
+     * @param integer $id
+     * @return mixed
+     */
+    public function actionView($id)
+    {
+        $model=$this->findModel($id);
+
+        if ($model->load(Yii::$app->request->post()) && $model->save()) {
+            if (!Yii::$app->user->can('/' . \common\components\helpers\Url::normalizeRoute('update'))) {
+                throw new HttpException(403);
+            }
+            Yii::$app->session->setFlash('kv-detail-success', 'Saved record successfully');
+            return $this->redirect(['view', 'id' => $model->id]);
+        } else {
+            return $this->render('view', ['model'=>$model]);
+        }
     }
 
     /**
@@ -53,7 +91,7 @@ class ArticleController extends Controller
         $model = new Article();
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['index']);
+            return $this->redirect(['index', 'returned'=>true]);
         } else {
             return $this->render('create', [
                 'model' => $model,
@@ -72,7 +110,7 @@ class ArticleController extends Controller
         $model = $this->findModel($id);
 
         if ($model->load(Yii::$app->request->post()) && $model->save()) {
-            return $this->redirect(['index']);
+            return $this->redirect(['index', 'returned'=>true]);
         } else {
             return $this->render('update', [
                 'model' => $model,
@@ -86,11 +124,31 @@ class ArticleController extends Controller
      * @param integer $id
      * @return mixed
      */
-    public function actionDelete($id)
-    {
-        $this->findModel($id)->delete();
-
-        return $this->redirect(['index']);
+    public function actionDelete($id) {
+        $post = Yii::$app->request->post();
+        if (Yii::$app->request->isAjax && isset($post['j-delete'])) {
+            if ($this->findModel($id)->delete()) {
+                echo Json::encode([
+                    'success' => true,
+                    'messages' => [
+                    'kv-detail-info' => 'Successfully deleted. <a href="' .
+                                                Url::to(['index']) . '" class="btn btn-sm btn-info">' .
+                        '<i class="glyphicon glyphicon-hand-right"></i>  Click here</a> to proceed.'
+                    ]
+                ]);
+            } else {
+                echo Json::encode([
+                    'success' => false,
+                    'messages' => [
+                        'kv-detail-error' => 'Cannot delete'
+                    ]
+                ]);
+            }
+            return;
+        } else {
+            $this->findModel($id)->delete();
+            $this->redirect(['index', 'returned'=>true]);
+        }
     }
 
     /**
